[이진트리] 04_07 코테 문제

+++ BackJoon/Class단계/4단계/16953번_A_B/소스.cpp
@@ -23,8 +23,8 @@ int main()
         pair<ll, int> front = q.front();
         q.pop();
 
-        int cur = front.first;
-        int cnt = front.second;
+        ll cur = front.first;
+        ll cnt = front.second;
 
         if (cur == b)
         {
